backend/requirements.txt
Django>=4.0.0,<4.1.0
djangorestframework>=3.13.0,<3.14.0
django-cors-headers>=3.11.0,<3.12.0
psycopg2-binary>=2.9.3,<2.10.0
dj-database-url==2.1.0
python-dotenv>=0.19.2,<0.20.0
djangorestframework-simplejwt>=5.0.0,<5.1.0
Pillow>=9.0.0,<9.1.0
django-filter==23.3
drf-yasg>=1.20.0,<1.21.0
gunicorn==21.2.0 